backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pillow==10.0.0
numpy==1.26.0
# optional OCR support:
pytesseract>=0.3.10
